+++ src/e2e/node-lib/project/src/lib/polynomial.ts
@@ -21,6 +21,14 @@ export class Polynomial {
     return this.normalize();
   }
 
+  toString(): string {
+    const expr: string = this.coefficients
+      .map((value: number, index: number): string => `(${value})x^${index}`)
+      .reverse()
+      .join(" + ");
+    return `P[${expr}]`;
+  }
+
   private normalize(): this {
     let i: number = this.coefficients.length;
     while (i > 0 && this.coefficients[i - 1] === 0) {
@@ -28,13 +36,5 @@ export class Polynomial {
     }
     this.coefficients.length = i;
     return this;
-  }
-
-  toString(): string {
-    const expr: string = this.coefficients
-      .map((value: number, index: number): string => `(${value})x^${index}`)
-      .reverse()
-      .join(" + ");
-    return `P[${expr}]`;
   }
 }